+++ src/components/ModalList.js
@@ -1,15 +1,17 @@
-import React, { PureComponent } from 'react';
+import React, {PureComponent} from 'react';
 import {
   FlatList,
   View,
   StyleSheet,
   TouchableHighlight,
   Text,
-  TouchableOpacity
+  TouchableOpacity,
+  Image,
 } from 'react-native';
-import { default as ModalBox } from 'react-native-modalbox';
+import {default as ModalBox} from 'react-native-modalbox';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import EventTracker from '../helper/EventTracker';
+import appConfig from 'app-config';
 
 class Modal extends PureComponent {
   state = {};
@@ -30,16 +32,28 @@ class Modal extends PureComponent {
     }
   };
 
-  renderItem({ item }) {
+  renderItem({item}) {
+    const isSelected = item.id === this.props.selectedItem.id;
+    const extraStyle = isSelected && styles.selectedItemContainer;
     return (
       <TouchableHighlight
         underlayColor="rgba(0,0,0,.1)"
-        onPress={() => this.props.onPressItem(item)}
-        style={styles.container}
-      >
-        <View style={styles.itemContainer}>
-          <Text style={styles.title}>{item.title}</Text>
-          <Text style={styles.description}>{item.description}</Text>
+        onPress={() => this.props.onPressItem(item, this.onClose)}
+        style={styles.container}>
+        <View style={[styles.itemContainer, extraStyle]}>
+          {!!item.image && (
+            <View style={styles.itemImageContainer}>
+              <Image style={styles.itemImage} source={{uri: item.image}} />
+            </View>
+          )}
+          <View style={styles.itemInfoContainer}>
+            <Text style={styles.title}>{item.title}</Text>
+            <Text style={styles.description}>{item.description}</Text>
+          </View>
+
+          {isSelected && (
+            <Icon name="dot-circle-o" style={styles.selectedIcon} />
+          )}
         </View>
       </TouchableHighlight>
     );
@@ -50,13 +64,12 @@ class Modal extends PureComponent {
       <ModalBox
         entry="bottom"
         position="bottom"
-        style={[styles.modal]}
+        style={[styles.modal, this.props.modalStyle]}
         backButtonClose
         ref={this.ref_modal}
         isOpen
         onClosed={this.props.onCloseModal}
-        useNativeDriver
-      >
+        useNativeDriver>
         <View style={styles.headingContainer}>
           <TouchableOpacity onPress={this.onClose} style={styles.iconContainer}>
             <Icon name="close" style={styles.icon} />
@@ -75,12 +88,12 @@ class Modal extends PureComponent {
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1
+    flex: 1,
   },
   modal: {
     height: '80%',
     borderTopLeftRadius: 15,
-    borderTopRightRadius: 15
+    borderTopRightRadius: 15,
   },
   iconContainer: {
     position: 'absolute',
@@ -88,42 +101,66 @@ const styles = StyleSheet.create({
     width: 30,
     height: 30,
     left: 15,
-    top: 15
+    top: 15,
   },
   icon: {
     fontSize: 22,
-    color: '#666'
+    color: '#666',
   },
   headingContainer: {
     padding: 30,
     borderBottomWidth: 1,
     borderStyle: 'solid',
-    borderBottomColor: '#ccc'
+    borderBottomColor: '#ccc',
   },
   heading: {
     fontSize: 30,
     fontWeight: '800',
     color: '#555',
     letterSpacing: 1.6,
-    textAlign: 'right'
+    textAlign: 'right',
   },
   itemContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
     width: '100%',
     padding: 15,
     borderBottomWidth: 0.5,
-    borderBottomColor: '#eee'
+    borderBottomColor: '#eee',
+  },
+  selectedItemContainer: {
+    backgroundColor: '#f5f5f5',
+  },
+  itemImageContainer: {
+    width: 55,
+    height: 55,
+    borderRadius: 8,
+    overflow: 'hidden',
+    marginRight: 15,
+  },
+  itemImage: {
+    width: '100%',
+    height: '100%',
+  },
+  itemInfoContainer: {
+    flex: 1,
   },
   title: {
     fontSize: 16,
     fontWeight: 'bold',
     color: '#444',
-    letterSpacing: 1.15
+    letterSpacing: 1.15,
   },
   description: {
     fontSize: 13,
     color: '#666',
-    marginTop: 2
-  }
+    marginTop: 2,
+  },
+  selectedIcon: {
+    fontSize: 20,
+    color: appConfig.colors.primary,
+    marginLeft: 15
+  },
 });
 
 export default Modal;