+++ src/containers/Launch/Launch.js
@@ -62,6 +62,9 @@ class Launch extends Component {
 
   handleAuthWithResponse = (response) => {
     const user = response.data || {};
+    const site = response.other_data?.site;
+    store.setStoreData(site);
+
     const {is_test_device} = user;
     const isTestDevice = this.handleTestDevice(is_test_device);