+++ src/helper/configKeyHandler/configKeyHandler.js
@@ -2,8 +2,9 @@ import store from 'app-store';
 
 export const isConfigActive = (key) => {
     const data = store.store_data;
+    
     if(data){
-        return !!data[key];
+        return !!Number(data[key]);
     }
     return false;
 }