+++ src/helper/configKeyHandler/types.js
@@ -1,4 +1,7 @@
 export const CONFIG_KEY = {
     SELECT_CITY_KEY: "choose_city_site_key",
-    SELECT_BIRTH_KEY: "choose_birth_site_key"
+    SELECT_BIRTH_KEY: "choose_birth_site_key",
+    SELECT_STORE_KEY: "choose_store_site_key",
+    
+    HIDE_PREMIUM_TAB_KEY: "hide_premium_tab_key"
 }